fixed msvc compil

+++ src/filters/dmx_nhnt.c
@@ -419,7 +419,7 @@ GF_Err nhntdmx_process(GF_Filter *filter)
 		gf_fseek(ctx->mdia, offset, SEEK_SET);
 
 		dst_pck = gf_filter_pck_new_alloc(ctx->opid, len, &data);
-		res = fread(data, 1, len, ctx->mdia);
+		res = (u32) fread(data, 1, len, ctx->mdia);
 		if (res != len) {
 			GF_LOG(GF_LOG_ERROR, GF_LOG_AUTHOR, ("[NHNT] Read failure, expecting %d bytes got %d", len, res));
 		}
